xspec: improve the typing of the get_xsabund routine
Since the return type depends on the arguments sent to the routine,
we can provide more-precise typing. Update to use Python 3.10
syntax (e.g. no need for Optional and Union).

It's not 100% clear from the typing documentation whether you should,
or need to, type the non-overloaded definition (I am reading proposed
changes to the documentation that doesn't make it clear); however, if
you don't set them then the annotations do not get set and so they
don't appear in help output. It also avoids an error reported by
pyright, so it seems worth doing.

Also add an example to the documentation for get_xsabund.

In looking at the XSPEC abundances, I noted that we did not document
the "felc" table, so do so for set_xsabund. In doing this I
re-arranged the list in the documentation so it's in alphabetical
order (ascending).