xspec: ensure set_xsabund("file") only after setting abundances
It turns out if you try to set the abundance setting to "file"
before reading in any data then you can crash XSPEC. So take
advantage of FunctionUtility::abundChanged to track this (as
this appears to be the reason for this method).

An alternative would be to ban "file" from the set_xsabund call,
but maybe people like switching between a file value and another
table (although there are other ways to do this in Sherpa).

I have also added some documentation to set_xsabund to explain
why we do not just call readNewAbundances but essentially
re-implement it (or, rather, now that we can use the C++ API
we keep the original code, as the xsFORTRAN interface did not
provide access to readNewAbundances). An alternative here
would be to drop the file-reading option from the C++ API and
make users read the data in Python and then use the
set_xsabundances routine to load in the data. This would
simplify the C++ code a bit *but*

- it would break existing code
- there is an advantage to having code that behaves as XSPEC
  does (in case there's subtle differences in reading in
  numbers and transferring them to the XSPEC internals)