xspec: rework set_xsstate
Previously set_xsstate required all the settings to be present before
it could work. This has its advantages, but does limit things as you
can't restore the state if Sherpa has added a new field. So, we now
just process those fields that are available and skip those that are
not.

As part of this work the internals to the XSPEC module have changed
to use a single dictionary for storing the data we need to record the
XSPEC state, rather than having multiple (two) variables for this.
This is in part because the current XSPEC interface that Sherpa uses
does not allow Sherpa to ask XSPEC what its state is, so the code has
to try and record these changes.