CI: add a WSL build
Add tests of building Sherpa on WSL using

- pip
- conda

The pip build is "everything but XSPEC and DS9" and the
conda build is "everything but DS9". Only a single python
version is targeted, as this is intended for monitoring
and not per-test validation.

This is based on the vedantmgoyal9/setup-wsl2@main action,
but significantly simplified by being able to use windows-2025,
which has wsl2 installed. Of note is the "wsl-run" script
that this action added, and has been copied here.

This attempts to use the existing architecture for running the
builds, but there are complexities in

- the build environment (e.g. what is installed in the Ubuntu
  environment used in WSL2 compared to the GitHub Linux runners)

- understanding the difference between Windows and Linux
  environment markers means that the settings are hard-coded to
  the steps rather than being set via env/matrix.