xspec: element access by table name
The XSPEC API lets us ask for an abundance value from any of the known
tables, not just the selected one. So, how can we provide this to the
user? I originally tried a very fine-grained set of routines -
e.g. providing access by name or by ion - but have settled on a much
simpler approach where

- get_xsabundances can be used with or without a table name
- get_xsabund is stuck using the selected abundance table

I just don't think that users of Sherpa need that many ways to
get the data, and that

    get_xsabundances(table)

should cover all their needs. That is, if they need access to a
particular ion they can extract the data themselves.

The code in sherpa/astro/xspec/src/_xspec.cc has been written
to make it easier to read - so we have the existing get_abund
routine and the new get_abund_from_table routine. The logic
could have been all placed in get_abund but

- it just complicates things that we don't need to complicate
- the sherpa.astro.xspec._xspec module is really an
  internal-to-Sherpa module, so we can add/remove code here as
  we want; it is the sherpa.astro.xspec interface we want to
  try and keep backwards compatible