xspec: provide access to the DEM data for plasma codes
The XSPEC API lets you get or set the DEM data. Here we chose just
to provide read access to the data. Unfortunately there's no
indication of what model it is for, so understandig the data is
very-much up to the user. There also doesn't appear to be a way
to track changes to this data - e.g. during a fit - which might be
nice to track but is not really going to fit in well with how the
API is set up and how Sherpa uses the models.

For the API I have gone for the simplest implementation I could
think of:

- there is no argument as there's no way to select any variant of
  the data
- the return is a pair (temperature, DEM) since there's no need to
  have individual accessor routines

It is not clear how best to use this data, so I have not tried
to add a "plot dem" option like XSPEC has - but I do show as an
example how to easily do this. The get_xsDEM() routine is currently
not exported by default, so users will need to know to explicitly
import it. I felt this was a good compromise as it's not clear how
we (and users) want to use this data.