Remove `_use_caching` from models.
Before this commit, caching happened if two things were true at the same time: The size of the cache is > 0 and `_use_caching == True`. Now there is only one source of truth: The setting of `model.cache`.

Since `_use_caching` was clearly marked as private, we can remove it without worrying about deprecation.

On the face of it, it looks like this will introduce difficulties with `ui.fit(cache=False)` which "runs a fit without caching". That might make a user think that a following run with `ui.fit(cache=True)` would then again use caching. That's not the case after this commit: `ui.fit(cache=False)` will set `cache=0` and will not reset it to a different value after the fit. We could do that in the `teardown` method, but what what we reset is to? To `cache=5`? What if a model should have permanently disabled for some very good reason? So, we would have to keep a temporary copy of the old `cache` value before the fit and then reset to that. That's complicated. And in fact, the old startup and teardown method also didn't do that.

Teardown used to have this line:
`self._use_caching = False`, so after a fit, `use_caching` was always False.

The following is with the Sherpa 4.17, showing that this behavior is pre-existing:
```
In [1]: from sherpa.astro import ui

In [2]: ui.load_arrays('data1', [1,2,3,4], [1,2,2,1.1])

In [3]: ui.set_model('data1', ui.box1d(name='boxmodel'))

In [4]: mdl = ui.get_model('data1')

In [5]: mdl._use_caching
Out[5]: True

In [6]: ui.fit(cache=False)

[...]

In [7]: mdl._use_caching
Out[7]: False
```

This was true even if the fit was done with `ui.fit(cache=True)`.

So, there is a change in behavior in this commit, but it's fixing this bug, so it's for the better.

You still disable the cache for all model if `ui.fit(cache=False)` and do not reset them after, but at least `ui.fit(cache=True)` no longer alters the size of the cache.

Update XSPEC model generation