tests: add a test of issue #2225 for load_user_stat
It is not clear if the documentation has got out of sync with
the code, or whether the code never really lived up to our
expectations. Add simple tests to check out whether we can use
the name of the user statistic to access the statistic.

As part of this add some related tests of the user-test
functionality.