sim: use get_thawed_pars to handle "extra" parameters - fix #2332
In Sherpa 4.16.1 we added the get_thawed_pars() method to models
that does what the existing code does but

a) it wraps it up
b) it ensures that we also include "extra" parameters that are free
   but not directly included in the model expression

I note that the thawedpars attribute internally uses
get_thawed_pars(), so most of this code was already using the
"correct" set of free parameters, so it is only a problem for
normal_sample when correlate=False.

This also fixes a small issue with error checking in
ParameterScaleVector (there's no issue number for this as it was
found during development of this code and is not a serious issue).

This also addresses some of the cases from #2337 but not all of
them.