Error out if model names are not strings
It is easy (#2222) to accidentally make a model where the name of a model is not a string, but another model instance. This might look like it works, but will lead to a useless model that is hard to debug for the average user because even printing out the model will fail in some cases.

This PR raises and error on model generation if the name is not a string.

fixes #2222