fit: code clean up
Some of these were suggested by 'ruff check' and are either

- use of isinstance rather than type to check for a value being
  of a particular type

- or, for places where isinstance is a "dangerous" check (thanks
  to the class hierarchy) use "type(x) is not y" rather than
  "type(x) != y"

There is also a minor code clean up where the results of np.where
are cached as it is used three times.

Other changes are suggested by type checkers - in particular
pyright - where a function parameter has type 'X | None' and then
later on there is code like

    if x is None:
        x = <make an X>

so at this point we know the value is not a None, but the type
checker does not. So, introduce a new variable which has type X,
e.g.

    if x is None:
        super_x = <make an X>
    else
        super_x = x

and then use super_x.