fit: improve typing
Try to add coverage to all the methods. It is not quite done,
and the type checkers don't like the code as there are some
issues over how data is combined (e.g. Data vs DataSimulFit)
and the iterated fit code works that need further work, but it
is an improvement.