tests: ensure corner-cases are checked (invalid arguments)
Ensure we check that we error out if sent invalid sigmarej
parameters. It turns out that there is an "error" in the
existing code, since it says things like

    raise SherpaErr("this argument is wrong")

but the resultant error actually says:

    SherpaErr("Generic Error")

This is thanks to the way the SherpaErr class is, and the
fact that the code that throws these errors should have been
using more-meaningful errors (i.e. subclasses of SherpaErr).
However, this is just a check of the existing code, so it does
not attempt to address this.