io: fix #2185
So, it turns out FITS files can use strings for the TLMIN/MAX
columns. Why, I don't know?

It looks like this is primarily a problem for the AstroPy backend,
since crates "seems to" handle this case (although that's because
"<value>" gets converted to 0 and that is what we care about for
the RMF case; for other files it could be more problematic but it
is actually hard to fix or workaround given how crates works).