tests: add tests of some problematic FITS behaviour
Issue #2185 and a recent CIAO helpdesk ticket have been about
what happens when sent an "interesting" FITS file:

- TLMIN/MAX values stored as strings
- PHA QUALITY column stored as a BYTE column

So, create example files and check we can read them in. It does
not seem worth adding the files to the sherpa test directory as
they are hopefully ephemeral (e.g. will get fixed in downstream
processing so we don't end up seeing them again). It is possible
that this code could be used in other tests, but let's worry
about abstracting this if we ever need to.

Now, the original test case for the TLMIN/MAX data came from
a RMF with a VLF column (F_CHAN), so it's important to check this
case (since it turns out the technique we end up using to fix
this has to special case the VLF handling). So the tests
check both scalar and VLF columns (PHA and RMF respectively)
which means the code has to deal with the FITS "heap", which was
a fun thing to have to learn about.