io: handle bit columns for PHA files with crates
A CIAO user has noted that NICER PHA files can't be loaded into
Sherpa, and it turns out that the QUALITY column has TFORM=B,
and this triggers some less-than-useful behaviour in crates. The
column gets read in as a 2D array with shape (ncols, 1), rather
than a 1D column of shape (ncols, ).

Unfortunately we can't directly know we are dealing with this,
but crates only uses np.uint8 when dealing with boolean data, so
we can at least handle this case. There is the possibility that
a user really wants a (n,1) column, but we are not writing a
generic data access layer here, so it should be okay.