io: update typing to Python 3.10 format
The Mapping and Sequence types are now taken from collections.abc,
Union types are labelled as a | b [ | .. ], and Optional[T] is now
T | None.

There are several minor lint changes (picked up by 'ruff check'):

- remove unused modules and symbols
- use 'foo is typeval' rather than 'foo == typeval'
  (as  didn't want to check whether we could use isinstance here,
   but not in all cases, such as those checking against object)

I also added a note to a section where we are assuming that the
F_CHAN and N_CHAN columns are set up the same (e.g. VLF or not),
but this assumption is a part of the OGIP standard.