BUILD: tweak conda installation environment
Change the location of the conda environment from within
the sherpa source directory - as given by the github.workspace
value - to its parent directory.

This turns out to be important if we change to a different
build system, such as meson-python [1]. It also seems to make
sense anyway; that is, make sure the condea environment is
separate from the Sherpa code.

I also apply this for the MacOS SDK location, for similar reasons.

Two questions are

a) is the parent directory a sensible location; I can't see why not
b) is there a better way to identify this location (or, some other
   location, following point a) as at present I just hard-code this
   into the yaml as "${{ github.workspace }}/../<value>"
   and also in setup_conda.sh for the macOS code (the macOS SDK)

---

[1] For example

https://github.com/rgommers/scipy/issues/82#issuecomment-974225733