TESTS: work around pytest assert rewrites failing with meson-python
I was getting worried about my understanding of things until I
found

https://github.com/mesonbuild/meson-python/issues/646

which explained why the "tests of the tests" fail here.

So, for now we try to fake the pytest assert and hope it's
somewhat similar (at some level any assertion message is
"nice to have" but I have been in cases where you get a failure
but no information on what actually failed and it is annoying).