tests: ensure the version matches
We have

    sherpa.__version__
    importlib.metadata.version("sherpa")

which should be the same, so check this.

However, editable installs present an issue, since the
sherpa.__version__ field will get a ".dirty" suffix as soon
as a git-tracked file is edited, but the metadata version
will not *if* it was installed before any edits were made
(even with no-build isolation, at least for setuptools).

This test was suggested by the 2024-10-07 version of

https://packaging.python.org/en/latest/discussions/single-source-version/

and added, in part, because experiments with switching to the
meson build system from setuptools showed it would be a useful
test. However, it may be too annoying given the editable issue,
so we may decide to revert this.