REGION: test CXCDM region parser, when available
The region code allows binary region files to be read in, but
ONLY if it was built with the USE_CXCDM_PARSER macro set. So we
can add a test or two of this behaviour, but how does the test
know what behaviour to expect? We could use a hueristic, such
as "does this appear to be a CIAO build", but it is **MUCH**
better to just encode the macro setting into the module so we
can ask it. We do this with the new

    sherpa.astro.utils._region.USE_CXCDM_PARSER

constant. It is 1 if it was set and 0 otherwise.

It appears that even though it was added with xxx_AddIntConstant
you can actually change this value. I do not see the point in
trying to deal with this as this is a highly-specialized variable
which users should not know about.