opt: fit statargs/statkwargs arguments are now unused
The fit method of the sherpa.optmethods.OptMethod class has support
for sending in extra arguments for the statfunc argument, so that
instead of accepting

    func(pars)

it can do

    func(pars, *statargs, **statkwargs)

This sounds like a sensible thing to allow, except that

- it is confusing to track down what is going on

  for example, we have no tests of this functionality, so is it
  actually useful or needed, and and it looks to have incomplete
  support (e.g. the error-estimation code have statargs/kwargs
  parameters but they are unused)

- it complicates the code (slightly)

- given that these extra arguments can not react to the parameter
  values / any looping that the optimization code does, there is
  no benefit to supporting this here, when a user who needs it
  can just wrap the statfunc argument with their own callback
  to add the information (this does assume that this is an uncommon
  task, but we have gone this far with Sherpa without it)

So we still allow the statargs and statkwargs to be set, but we do not
do anything with them, and mark them as having the type "Any", as
well as ensuring their default vaule is "None".

This means that we can avoid wrapping statfunc in a callback before
sending it to the optimizing function.