opt: replace a local lambda with the InfinitePotential class
Move a local lambda to a class, defined in the module, to allow
the code to run when a non-fork method is selected for
multiprocessing.

The code implements an "infinite" potential - that is, if any
parameter values are outside the valid ranges then the "max"
value (for float64) is returned. There is already an existing class
that implements this, sherpa.optmethods.optfcts.InfinitePotential,
but it is subtly different:

1. the return value is different (a float rather than
   tuple[float, np.ndarray]),

2. and the limits check do not check for NaN values
   (this could be easily added, but it does not seem worth
   it given point 1)

The current approach is to use the same symbol ("InfinitePotential")
and rely on the namespace (that is, which module the code is in) to
define the behaviour. This may turn out to be a bad idea.

The code used to allow sending in multiple arguments, but we never
made use of the capability, so it now requires the function being
wrapped to take a single ndarray argument (not ArrayType since the
code assumes you can say "np.any(x < ...)" rather than "ArrayType or
None".