opt: replace a local lambda with the InfinitePotential class
However, this is **not** exactly InfinitePotential from
sherpa.optmethods.optficts, because

1. the function returns a float and not tuple[float, np.ndarray]
2. the limits check do not check for NaN values
3. this routine can accept 1 or more arguments, not just
   1 [do we actually use this though?]

The latter could be changed - i.e. the code could add a "is not
NaN" check - but the former is still a difference that it is not
clear how best to handle. So the current approach is to use the
same symbol ("InfinitePotential") and rely on the namespace (that
is, which module the code is in) to define the behaviour. This
may turn out to be a bad idea.