opt: replace a local lambda with the InfinitePotential class  WIP
However, this is **not** exactly InfinitePotential from
sherpa.optmethods.optfcts, because

1. the function returns a float and not tuple[float, np.ndarray],

2. and the limits check do not check for NaN values
   (this could be easily added, but it does not seem worth
   it given point 1)

The code used to allow sending in multiple arguments, but we never
made use of this so it now requires the function being wrapped to take
a single ndarray argument (not ArrayType since the code assumes you
can say "np.any(x < ...)".

The current approach is to use the same symbol ("InfinitePotential")
and rely on the namespace (that is, which module the code is in) to
define the behaviour. This may turn out to be a bad idea.