estmethods: make most arguments to compute keyword-only
This is technically a style change, but it makes it more-obvious
if an argument has been placed in the wrong location, or is just
missing.

A note has been added to point out that the statargs and statkwargs
arguments are currently unused. They could be removed, but leave
in for future work.