estmethods: compute statargs/statkwargs arguments are unused
The statargs and startkwargs arguments for the compute methods are
currently unused, so set their default values to None to make it clear
that there's no information provided with them. It means that types do
not have to be given, and removes a "mutable argument" warning from
pylint.

The code will display a warning (using the Sherpa logger) if they
are set to anything but None to point this out.