statargs/kwargs arguments for statistic calls are now unused
Adding types for the various optimization-related code has shown
partial support for sending in extra arguments to routines that
calculate the statistic value for a set of parameters. Examples
include that the estmethods code allows both statargs and statkwargs
to be set but does nothing with this information, and some code
allows for extra positional arguments but not named ones.

This makes following the code, and adding types, harder than it
could be, in part because the code is written in a functional
style (in that functions are passed around) and there are a lot
of places where "adapters" are needed to support different
calling styles (or extracting out a subset of the return data).

So these changes

- remove passing in extra arguments to these callback-like
  functions

- changes the default values for the statargs and statkwargs to
  None

  (recent work actually did this for most/maybe all of the
   statkwargs parameters, since they used to be set to a default
   of '{}' and pylint complained about using a mutable argument,
   e.g. PR #1778)

- creates a WARNING-level log message if the routines are ever
  called with statargs or statkwargs set, to let users know this
  information is not being used

Now, if users do need to send in extra arguments then it is
suggested that they create a new function that adds in the extra
data (e.g. follows the callback-like approach used by these
routines).

Note that there were are tests that exercised the "send in extra
information with statargs or statkwargs" functionality.