utils: very-minor typing additions
Label several routines as returning a bool (this can turn out
to be problematic, since numpy values return a np.bool_, so
ideally these routines should explictly case the return to
avoid this, but leave that for later work).

A comment about possibly returning a TypeGuard has been removed
because it wasn't helpful: yes, the return could return a TypeGuard,
but the suggested TypeGuard value (of a Sequence) is not obviously
the correct "thing" (should it be Sequence or Iterable or ...).