types: better typing of StatsFunc/add OptFunc
It is hard to track down exactly what the type of the statistics
function is: the code only provides a single argument version,
but it allows for extra arguments (passed around via statargs and
statkwargs arguments). Previously this was represented as

    Callable[..., StatResults]

but, as we know the first argument, we can say

    Callable[Concatenate[ArrayType, P], StatResults]

where P is a ParamSpec object (needed for Python 3.10).

An alternative would be to just drop the statargs/kwargs support.

With Concatenate we can make a reasonable type for the optimization
functions. An alternative would be to use a Protocol, but the
API does not seem worth doing this for (as the only arguments we
can give types to here are named arguments).