BUILD: allow Sherpa to build without FFTW
If the FFTW library is not available we can not build the
sherpa.utils._psf module, but we can still do quite a lot. This commit
makes the FFTW library be an optional extra rather than required. The
main changes are

a) add the disable-fftw build action (in setup.cfg)
b) warning users when the _psf module is not available
c) marking those tests as optional that require the _psf module

In general this is as expected, but there are a couple of image tests
that have been shown to require the _psf module even though they do
not actually do any PSF convolution. This is because they require the
"bounding box" routines. We could look at separating this from the PSF
code if we felt it worthwhile. However, this is probably not worth it
as these code paths require other optional build elements (e.g. the
region code), which suggests it is not often users would hit this
condition.