fit: drop "iterative fit method" in output when not set
This was found by a type checker (after some other changes, not
included here). The "error" here is the check

    if val is not None or val != 'none':
        add-output including val.capitalize()

The type checker noted that a None value does not have a
capitalize method (although technically it does). The change is
to use "and" rather than "or" - which also makes it match the
ErrorEstResults code match the FitResults code.

This then leads to a test failure, since we no-longer create
the line

    Iterative Fit Method  = None

in the output.

This is technically a breaking change, since the output of
UI commands like `conf()` will change, but it's hard to easily
document this, and it is unlikely that most users would notice.