tests: rework the estmethods tests
Since most of tests all call the compute method the same way,
introduce the get_working_kwargs routine to return a kwargs
dictionary that can be used to essentially set all the arguments
in one call. This is then used to

a) ensure the arguments are called correctly, since it uses calls
   arguments by name rather than position

   This shows a logical error in the first test (it was
   test_covar_failures1), which was actually failing because the call
   was missing the get_par_name argument, but it looks like the test
   should have been to check it fails when the statfunc argument
   was None. Actually, this is supposition on my part, since there are
   no comments in the test to indicate what it was meant to be
   testing, but what was test_projection_failures1 behaves this way,
   and it is a more-meaningful test.

b) make it easy to swap out one of the arguments for an invalid
   value

A number of tests check that the Covariance and Projection classes
behave the same way (that is, error out for certain invalid argument
settings). The code has been reqorked to allow the tests to be
parameterized by the class, which

- avoids duplication
- allows adding tests of the Confidence class

It turns out that Confidence errors out in different ways to the
other classes, which complicates the tests slightly. There is also
the fact that Covariance can accept None for the fitfunc but the
other classes can not.

Adding the Confidence classes changes the behaviour of the
get_par_name callback, since it was never called with the other
classes but is now. Since the behaviour of this routine is not
being tested, remove the recently-added assert and replace with
the original pass call.

The code could be updated to add checks to ensure they handle these
error cases in the same manner, but that is a future possibility.