tests: add gridsearch test that sets the method argument
The method argument in the gridsearch code is not tested, so add
a **very basic** test that at least exercises it. This is meant
as a regression test, and is **not** a test that changing the
method argument is doing the correct thing (since I do not have
the time to dig into this code to try and work out what a sensible
set of tests would be).