opt: clean up of method argument handling in gridsearch
So, rather than looking for "expected" variants of the method
name - like "neldermead", "NelderMead", ... - just convert to
lower case. This technically allows more method arguments to
match, but we have this logic in other parts of the system.

Add a warning if the method is unknown.

We do not have good tests of what setting the method does, but
that has been true ever since this code was orignially added to
Sherpa.