fit: drop support for statargs/statkwargs arguments
The fit and sigmarej routines still accept statargs and statkwargs
arguments but they are ignored. It was not clear if the existing
functionality worked as there were no tests and tracking the
information flow is awkward given the convoluted paths here.

The suggestion is that the user should define the callback routine to
apply any extra information, rather than it being done by Sherpa.  If
this turns out to be a useful and common operation we can come back
and revisit this.