opt: change SimplexBase typing rules
The code has been written assuming that the seed and factor
arguments were always given to the SimplexBase-based classes,
but it turns out ncoresnm.py can set these both to None.

So, update the typing rules to point this out.

For the seed value it is okay (I think) since the seed value is sent
to np.random.seed() and that allows a None value. There is the
possibility it is used elsewhere, but I don't think so (the code is
not clear to follow and there's no commentary or documentation to
guide the reader).

For the factor argument it is actually a bit surprising that factor
can be None, but it turns out that it is okay for this particular use
case because the factor value is unused. The usability of the API here
is poor, as it is not-at-all clear that this is the case. However,
this is all low-level code which we want to avoid changing too much,
too quickly. So a comment has been added where factor is set to None,
just to guide the reader.