opt: use rng rather than seed for SimplexBase
In #1735 there was a push to move the random-number-using code
away from global state (calling random.seed/np.random.seed) and
towards using local state (the NumPy random generator classes).

The SimplexBase class from sherpa.optmethods.opt did not fully change
over, so do so here. Previously the class could be sent a seed value
and a rng argument (the random generator) but it was okay for rng to
be None. The new code still allows this for the initialization, but
internally now creates a RNG object and then this gets used, rather
than using the global state.

As part of this the seed argument is now "deprecated" - at least for
the non-initialization calls - as it is no-longer used. To support a
future change where the argument is removed the initialization and
seed-using calls have been changed so that a number (maybe all) of the
arguments must be given by name.