opt: remove seed argument from SimplexBase
As the seed argument is no-longer used, remove it to make it clearer
how the RNG is set. This is a breaking change, but this is for
low-level code that is only used by the optimizer code, and is
undocumented (the assumption being that no-one is using it outside of
Sherpa).