TESTS: update SimplexBase related tests
PR #1735 updated the Sherpa code base to allow a RNG object to be sent
around, rather than relying on global state. The tests did not fully
test the behaviour of the SimplexBase objects: the initialization can
be sent seed, rng, or both and we want to check we get the same back
however it is sent (modulo the different values, as seed in int | None
and rng is RNG-class | None).

So, update a few tests just to check we cover all three cases:

 - seed set, rng is None
 - seed set, rng set
 - seed not set, rng set

As this is a low-level class we do not have tests of failure cases,
like seed and rng not being set.