opt: SimplexBase now uses latest NumPy RNG
Rather than use the RNG that was used with numpy.random.seed and
numpy.random.RandomState, use the default RNG - that is
numpy.random.defaut_rng - to set up the rng used by the SimplexBase
classes.

This does change the random numbers generated by a given seed, and so
several of the tests have had to be updated. I am assuming that these
new values are "okay", but really these tests were written as
regression tests rather than functional ones.