opt: SimplexBase now uses latest NumPy RNG
Rather than use the RNG that was used with numpy.random.seed and
numpy.random.RandomState, use the default RNG - that is
numpy.random.defaut_rng - to set up the rng used by the SimplexBase
classes.

This does change the random numbers generated by a given seed, and so
several of the tests have had to be updated. I am assuming that these
new values are "okay", but really these tests were written as
regression tests rather than functional ones.

There are changes to the output when moncar is run with numcores > 1;
for numcores=2 and 3 the nfev is now the same when it was different.
Is this just a change in the RNG or something more complex? With these
tests, which are intended as regression tests, it is hard to tell.