opt: internal clean up
Import numpy as np.

Use super() rather than refer to the superclass by name.

Clean up a __init__ call that used a list as the default value.  The
code now uses None to mean the old list-value should be used, which
reduces the chances of accidentally changing future calls to the
initializer. The subclass initializer has been removed as it is just
the same: previously it had been left in in case the default options
were going to be changed, but it is cleaner if we do not leave it
in (so that if the default list of strategies is to be changed it is
obvious from the code).