opt: add typing rules to ncoresnm/de
ncoresnm:

The typing of the finalsimplex argument is "interesting" as it is
not clear where it can be

- None
- int
- sequence of int

At the moment the typing is generally "int" or "int | None".

The multi-core code needs improved typing statements, as do many of
the places where a function is labelled as Callable rather than giving
it a more-specific type.

ncoresde:

The typing rules assume input is a ndarray rather than the parent
ArrayType argument since there are some places where the code
multiplies a scalar by an array. Ideally the code would be updated to
convert to a ndarray where appropriate, but that is for later.

opt:

Introduce types used by the above.