opt: internal rework of multi-core code
The multi-core code here does not quite match the versions in
sherpa.utils.parallel, which is why they do not use those functions.
However, the code can be written to better use the parallel module: in
this case the code was confusing to follow since the run_tasks routine
was written assuming it would be sent tasks, each of which contained
multiple tasks, but here wach task is a single job, so the "result
expansion" was hard to follow and even harder to type. So tweak the
code to make the results of the calc call easier to type, and so
easier to process.

There may be more changes, since it would make sense to shard the
processes by the number of cores rather than assuming that each task
can be run separately.