opt: internal clean-up of ncoresnm/de
Following this code is hard since the signatures of routines change
in subclasses. Normally this is just a positional argument changing
name (e.g. "xpar" to "x"), but at least one class changes the
call method, swapping the order of the parameters to its superclass.
In this particular case the change in order did not lead to any
problems, since it was called with the swapped ordering, but it is
tricky to follow as is. It also makes it hard to write typing
statements.

So we have a breaking change here, but it is in very low-level code,
so I have added documentation notes (and any code using the old
form will fall over rather than return the wrong results).

So this commit basically normalizes a number of calls to match their
superclass, forces some calls to use named arguments for some (or all)
of their arguments (mainly to just make the code document what it's
doing, and to avoid possible surprises), and some minor code clean ups
like:

- use a new variable when potentially type changing a value
  (this is only done for some cases of changing an ArrayType to
  a np.ndarray),

- ensure we return a ndarray and not a list for a call (this call is
  actually unused and should probably be made to raise
  NotImplementedError, but it does act to document the intent of the
  class),

- replace "if <truthy> then return True; return False" with
  "return bool(<truthy>)",

- add comments for classes that are only used in the tests (these
  are specialized tests that require running the code directly, and
  so are not run in CI or via pytest), as some of the code contains
  obvious bugs which I keep on noticing (I do not want to remove this
  code just yet),

- and remove a commented-out print statement.