opt: add typing rules to opt and ncoresnm/de
It is now possible to improve the typing of a number of callable
arguments. For example, the multi-core code saw a number of Callable
arguments changed to more-specific types.

The use of the ArrayType type points out some places where the code
assumes the input is a ndarray, so there is a balance between using
ArrayType and ndarray. One question is whether we have this balance
correct?

It is unclear over whether the finalsimplex argument can be

- None
- int
- sequence of int

At the moment the typing is generally "int" or "int | None" even
though there is code to support sending in a sequence of int. Fixing
this is left for later work.

There are also questions whether some arguments can be "x | None",
such as maxfev where it may be int or int | None. At the moment the
typing is best-effort.

Numeric scalar values are generally being represented as float, even
when it may also be a np.float64 (or np.float32) value, which pyright
and mypy don't treat as float. There is an attempt to ensure
conversion to a Python float happens (from the NumPy one) to avoid
complaints from the type checker, but it is not guaranteed to catch
everything.